added dokumentation and handling to protect deletion of userroles. Functionality supported from backend but not used in frontend

+++ backend/src/main/java/at/fhj/swengs/delorian/service/UserRoleService.java
@@ -5,10 +5,7 @@ import at.fhj.swengs.delorian.repository.UserRoleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 
 @Service()
 public class UserRoleService {
@@ -28,8 +25,8 @@ public class UserRoleService {
         return userRoleRepository.findAll();
     }
 
-    public Optional<Role> findByRoleName(String username) {
-        return userRoleRepository.findById(username);
+    public Optional<Role> findByRoleName(String rolename) {
+        return userRoleRepository.findById(rolename);
     }
 
     public Role save(Role entity) {
@@ -37,6 +34,25 @@ public class UserRoleService {
     }
 
     public void delete(String rolename) {
-        userRoleRepository.deleteById(rolename);
+        // UserRoles hard coded used in SecurityConfig for JWT - never delete them!
+        // Think twice to delete userRoles and use this feature whisely!
+        List<String> protectedRoles = new ArrayList<String>() {{
+            add("ROLE_ADMIN");
+            add("ROLE_PRJMGR");
+            add("ROLE_USER");
+        }};
+        if(protectedRoles.contains(rolename)) {
+            // Do not delete userRoles which are used hard coded for backend Security!
+        } else {
+            Optional<Role> optEntity = findByRoleName(rolename);
+            if (optEntity.isPresent()) {
+                Role entity = optEntity.get();
+                if (entity.getAssignedUsers().isEmpty()) {
+                    userRoleRepository.deleteById(rolename);
+                } else {
+
+                }
+            }
+        }
     }
 }
